update documentation to next nef features
