+++ Documentation.app/Contents/MacOS/Command Line Tool.playground/Pages/Generating a Markdown document.xcplaygroundpage/Contents.swift
@@ -1,0 +1,39 @@
+// nef:begin:header
+/*
+ layout: docs
+ title: Generating a Markdown document
+ */
+// nef:end
+
+// nef:begin:hidden
+import Bow
+Nef.Playground.needsIndefiniteExecution(false)
+// nef:end
+
+/*:
+ ## 🔖 Generating a Markdown document
+ Xcode Playgrounds let you write comments in Markdown format using the symbols `//:` for single line comments, or `/•: ... •/` for multiline comments. Inside this comments, you can use any Markdown syntax; an exmaple:
+ 
+ ```swift
+ /•:
+ ‌‌ ### This is a heading 1
+ This is regular text. *This is bold text*. [This is a link](http://bow-swift.io).
+ •/
+ 
+ protocol MyProtocol {}
+ 
+ //:‌‌ ## This is a single line heading 2‌‌
+ ```
+ 
+ It makes Xcode Playgrounds the proper tool to write an article with compilable examples. The command provided by `nef` to generate the Markdown files is:
+ 
+ ```bash
+ ➜ nef markdown --project <nef playground> --output <path>
+ ```
+ 
+ Options:
+ 
+ - `--project`: Path to the folder containing the Xcode project with Xcode Playgrounds.
+ - `--output`: Path where the resulting Markdown project will be generated.
+ 
+ */
