+++ Documentation.app/Contents/MacOS/Command Line Tool.playground/contents.xcplayground
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<playground version='6.0' target-platform='ios' display-mode='raw'>
+    <pages>
+        <page name='Creating a nef Playground'/>
+        <page name='Compiling a nef Playground'/>
+        <page name='Creating a Playground Book for iPad'/>
+        <page name='Generating a Markdown document'/>
+        <page name='Jekyll integration'/>
+        <page name='Exporting Carbon code snippets'/>
+    </pages>
+</playground>
