+++ Documentation.app/Contents/MacOS/First Steps.playground/contents.xcplayground
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <playground version='6.0' target-platform='ios' display-mode='raw'>
     <pages>
-        <page name='Getting started'/>
-        <page name='Resources'/>
+        <page name='How to use carbon api?'/>
+        <page name='How to export my snippets?'/>
     </pages>
 </playground>
