--- Documentation.app/Contents/MacOS/Generation.playground/contents.xcplayground
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<playground version='6.0' target-platform='ios' display-mode='raw'/>
