--- Documentation.app/Contents/MacOS/Integration.playground/contents.xcplayground
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<playground version='6.0' target-platform='ios' display-mode='raw'>
-    <pages>
-        <page name='Jekyll'/>
-        <page name='Carbon'/>
-    </pages>
-</playground>
