+++ Documentation.app/Contents/MacOS/Library.playground/Pages/Modules.xcplaygroundpage/Contents.swift
@@ -1,0 +1,47 @@
+// nef:begin:header
+/*
+ layout: docs
+ title: Modules
+ */
+// nef:end
+
+/*:
+ ## Modules
+
+ Provides you the whole nef features into your macOS project. It enables you to build new tooling on top of nef. nef API gives you an easy access to render Markdown files, Jekyll content and Carbon images in a functional way. Also you can compile, clean your nef Playrounds, even make an Xcode Playground/Playground Book compatible with 3rd-party libraries.
+
+ ### Compile
+
+ You can find more information in [Compile API docs]()
+
+ ### Clean
+
+ You can find more information in [Clean API docs]()
+
+ ### Playground
+
+ You can find more information in [Playground API docs]().
+
+ ### Markdown
+
+If you need to render the content of a Playground page, or anything else that combine [**markup formatting**](https://developer.apple.com/library/archive/documentation/Xcode/Reference/xcode_markup_formatting_ref/index.html) and code.
+
+ You can find more information in [Markdown API docs]().
+
+ ### Jekyll
+
+ When you need to render content that combines Markdown, code, and [**nef commands**](https://github.com/bow-swift/nef#-generating-markdown-files-for-jekyll) with the Jekyll style.
+
+ You can find more information in [Jekyll API docs]().
+
+ ### Carbon
+
+ If you want to share some Swift snippets. Carbon is a cool tool for this, and nef nicely integrates with it. nef lets you render Carbon images given a configuration (including code to render and style)
+
+ You can find more information in [Carbon API docs]().
+
+ ### Playground Book
+
+ You can find more information in [Playground Book API docs]().
+
+ */
