+++ Documentation.app/Contents/MacOS/Podfile
@@ -2,11 +2,11 @@ target 'Documentation' do
   platform :ios, 12.0
   use_frameworks!
 
-  pod "Bow", "~> 0.6.0"
-  pod "BowEffects", "~> 0.6.0"
-  pod "BowRx", "~> 0.6.0"
-  pod "BowOptics", "~> 0.6.0"
-  pod "BowGeneric", "~> 0.6.0"
-  pod "BowRecursionSchemes", "~> 0.6.0"
-  pod "BowFree", "~> 0.6.0"
+  pod "Bow", "~> 0.7.0"
+  pod "BowEffects", "~> 0.7.0"
+  pod "BowRx", "~> 0.7.0"
+  pod "BowOptics", "~> 0.7.0"
+  pod "BowGeneric", "~> 0.7.0"
+  pod "BowRecursionSchemes", "~> 0.7.0"
+  pod "BowFree", "~> 0.7.0"
 end
