+++ Documentation.app/Contents/MacOS/Quick start.playground/Pages/Getting started.xcplaygroundpage/Contents.swift
@@ -1,0 +1,44 @@
+// nef:begin:header
+/*
+ layout: docs
+ title: Getting started
+ */
+// nef:end
+
+/*:
+ ## Getting started
+ `nef` is a command line tool to ease the creation of documentation in the form of Swift Playgrounds. It provides compile-time verification of documentation and exports it in Markdown format that can be consumed by [Jekyll](https://jekyllrb.com/) to generate websites.
+
+ ### 💻 Installation
+
+ #### 📟 Using [Homebrew](https://github.com/bow-swift/homebrew-nef) (preferred)
+
+ ```bash
+ ➜ brew install nef
+ ```
+
+ > It will warn you if there is a missing dependency and will provide guidance to install it.
+
+ 📣 You will find more information in [Command Line Tool](/docs/command-line-tool/creating-a-nef-playground/) section.
+
+ &nbsp;
+
+ #### 📦 Using [Swift Package Manager](https://developer.apple.com/documentation/xcode/creating_a_swift_package_with_xcode)
+
+ `nef` can be consumed as a library in your **macOS project**.
+
+ ```swift
+ .package(url: "https://github.com/bow-swift/nef.git", from: "{version}")
+ ```
+
+ It is an excellent option if you want to use all nef features in your macOS app, even to build new tooling on top of nef.
+
+ 📣 You will find more information in [Library](/docs/command-line-tool/creating-a-nef-playground//) section.
+ 
+ &nbsp;
+
+ #### 🔌 Using [Xcode Editor Extension](https://github.com/bow-swift/nef-plugin)
+
+ Some of `nef` features can be used directly in Xcode as an Extension. You can install it directly from [**App Store**](https://apps.apple.com/app/nef/id1479391704?mt=8) or downloading the last binary from the [**releases section**](https://github.com/bow-swift/nef-plugin/releases).
+ 
+ */
