--- Documentation.app/Contents/MacOS/Quick start.playground/Pages/Installation.xcplaygroundpage/Contents.swift
@@ -1,43 +1,0 @@
-// nef:begin:header
-/*
- layout: docs
- title: Installation
- */
-// nef:end
-
-/*:
- ## 📥 Installation
-
- 📟 **Using** [**Homebrew**](https://github.com/bow-swift/homebrew-nef)
-
- > It will warn you if there is a missing dependency and will provide guidance to install it.
-
- ```bash
- ➜ brew tap bow-swift/nef
- ➜ brew install nef
- ```
-
- You can upgrade the version of `nef` using the following command
-
- ```bash
- ➜ brew upgrade nef
- ```
-
- &nbsp;
-
- 📦 **Using** [**Swift Package Manager**](https://developer.apple.com/documentation/xcode/creating_a_swift_package_with_xcode)
-
- `nef` can be consumed as a library in your **macOS project**.
-
- ```swift
- .package(url: "https://github.com/bow-swift/nef.git", from: "{version}")
- ```
-
-  &nbsp;
-
-
- 🔌 **Using** [**Xcode Editor Extension**](https://github.com/bow-swift/nef-plugin)
-
- Some of the `nef` features can be used directly in Xcode as an Extension. You can install it directly from [**App Store**](https://apps.apple.com/app/nef/id1479391704?mt=8).
-
- */
