--- Documentation.app/Contents/MacOS/Tutorials.playground/contents.xcplayground
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<playground version='6.0' target-platform='ios' display-mode='raw'>
-    <pages>
-        <page name='How to use nef library?'/>
-        <page name='First Steps with Carbon'/>
-    </pages>
-</playground>
