+++ docs/_data/features.yml
@@ -1,16 +1,16 @@
 content:
     - title: Verification
-      description: Eases the creation of Xcode Playgrounds with support for 3rd-party libraries.
+      description: Eases the creation of Playgrounds with support for 3rd-party libraries.
       icon: img/main-image-primary.svg
 
     - title: Compile
-      description: Compiles Xcode Playgrounds with support for third party libraries from the command line.
+      description: Compiles Playgrounds with support for third party libraries from the command line.
       icon: img/main-image-secondary.svg
 
     - title: Generation
-      description: Generates Playground Books from Swift packages and Markdown files from Xcode Playground.
+      description: Generates Playground Books from Swift packages and Markdown files from nef Playground.
       icon: img/main-image-tertiary.svg
 
     - title: Integration
-      description: Integrates Xcode Playgrounds with Carbon to export code snippets and Jekyll to create a microsite.
+      description: Integrates Playgrounds with Carbon to export code snippets and Jekyll to create a microsite.
       icon: img/main-image-quaternary.svg
